backend/requirements.txt
Flask
python-dotenv
google-api-python-client
google-auth
google-cloud-speech
openai
airtable-python-wrapper
google-cloud-storage
pytube
anthropic
pydub
google-auth-oauthlib 
google-api-python-client
crewai